+++ index.js
@@ -1,14 +1,10 @@
+// This is the handler for Lambda.
+// Processes on Guild per run.
+
 const GuildIterator = require('./src/GuildIterator.js')
 
-const main = async function(){
+exports.handler = async (event) => {
   const gi = new GuildIterator('manifest.json')
   await gi.fetchManifest()
   await gi.processNextGuild()
 }
-
-main()
-  .then(() => process.exit(0))
-  .catch(error => {
-    console.error(error);
-    process.exit(1);
-  });
